Move Query#toStatementStats to a utility class
